op-deployer: Add intent builder (#15136)
* op-deployer: Add intent builder

Based on conversations with @protolambda, we need an easier way of defining chain deployments for use within tests. This PR does that by introducing a new "builder" API for op-deployer intents. This API allows users to define chain intents using a more functional style and with fewer required fields. In the future, this will allow chain intents/state to be automatically hydrated from existing on-chain data to allow for incremental deployments.

For the time being, the builder exists in a dedicated subpackage of op-deployer. we can change this later. I also added support for customizing the start timestamp/block hash of L2s rather than relying on `time.Now()` or the `latest` block. This will make test more deterministic.

* CRR updates

* fix test