test(sysext): add support for control plane (#15461)
This implements the control plane interface for the sysext side of devstack.
Note that only kurtosis devnets provide a corresponding implementations for now.

For convenience, the orchestrator now embeds a DevnetEnv, which is the
enriched wrapper around descriptors.DevnetEnvironment (data-only
object)

Validated using a slightly adjusted version of
devnet-sdk/devstack/sysgo/control_plane_test.go but that's a temporary
measure. A real test against the devstack interfaces will be required.