op-challenger/op-dispute-mon: Load dispute game factory from multiple network args (#15744)
* op-challenger: Support loading factory address with multiple networks

Validates that the DisputeGameFactoryProxy address for all specified chains matches. Can still be overridden with --game-factory-address if required.

* op-dispute-mon: Support loading factory address with multiple networks

Validates that the DisputeGameFactoryProxy address for all specified chains matches. Can still be overridden with --game-factory-address if required.
Increases safety of monitoring by allowing operators to name all networks they expect to be monitoring via the shared factory and dispute-mon will fail to start if one of them isn't actaully sharing that factory (according to the superchain-registry).